% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCermine.R
\name{pdf_to_convert}
\alias{pdf_to_convert}
\title{Run cermine on a PdF.}
\usage{
pdf_to_convert(directory)
}
\arguments{
\item{directory}{a directory containing the the PDF to convert.}
}
\value{
A vector containing the file reference to the JATS xml file.
}
\description{
Reads a PDF and converts it's content to a Journal Article Tag Suite (JATS) xml file.
}
\examples{

RunCermine(directory)

}
\author{
Jason Mumbulla, \email{jasonmumbulla@gmail.com}
}
